Added generic workdir path resolver
Implemented a workdir interface, to create work directories,
in parallel of the data paths interface.

Implemented support in orm utility for mutable src and sysroot,
Now resolves srcdir using git-rev parse --show-toplevel to find
repository if needed. Complemented with support for environment
variables if these defaults are not good enough.

* bconf.mk: Support for new workdir interface.
* include/orm.h: Added workdir interface.
* liborm/: Moved in lib/
* lib/workdir.c: New workdir interface.
* src/orm.c: See above for more.