Implementation of Lindworm utility
Created lindworm, which creates packages from the content
of the destdir or objdir of the sandbox. These sandbox directories
are neither accessible nor persistent. Lindworm is meant to be used
on working bsys and toolchains, and run a clean build from scratch.
Use orm if you want iterative builds and caching instead.

Also factorized the function creating a path from a process' standard
output. Not put in the orm library, as it is not really orm-related.
Added configuration build option to specify the default srcdir command.

* bconf: New default srcdir command option.
* bconf.mk: Support for lndworm binary, cmdpath and new srcdir command option.
* src/cmdpath.c: Factorized path-from-stdout function.
* src/lndworm.c: Lindworm utility preliminary version.
* src/orm.c: cmdpath + default srcdir command support.